test(withstyleobject): add combined tests
