+++ __tests__/serializers/slateToHtml/index.spec.ts
@@ -1,8 +1,5 @@
-import { ChildNode, Element } from 'domhandler'
+import { Element } from 'domhandler'
 import { slateToHtml, slateToDomConfig } from '../../../src'
-
-const postcss = require('postcss')
-const postcssJs = require('postcss-js')
 
 describe('slateToHtml expected behaviour', () => {
   it('encodes HTML entities', () => {
@@ -312,154 +309,6 @@ describe('custom config', () => {
         style: ({ node }: { node?: any }) => {
           return new Element('span', {
             ...(node.style?.fontSize && { style: `font-size:${node.style.fontSize};` }),
-          })
-        },
-      },
-    }
-    expect(slateToHtml(slate, config)).toEqual(html)
-  })
-})
-
-describe('style attribute css transforms with postcss', () => {
-  const transformStyleObjectToString = (style: { [key: string]: any }) => {
-    const postcssOptions = {
-      parser: postcssJs,
-      from: undefined,
-    }
-    return postcss()
-      .process(style, postcssOptions)
-      .css.replace(/(\r\n|\n|\r)/gm, ' ')
-      .replace(/\s\s+/g, ' ')
-  }
-
-  it('element transforms', () => {
-    const html =
-      '<p style="font-size: 96px; --text-color: #DD3A0A; @media screen { z-index: 1; color: var(--text-color) }"><strong>Paragraph</strong></p>'
-    const slate = [
-      {
-        type: 'p',
-        style: {
-          fontSize: '96px',
-          '--text-color': '#DD3A0A',
-          '@media screen': {
-            zIndex: '1',
-            color: 'var(--text-color)',
-          },
-        },
-        children: [
-          {
-            bold: true,
-            text: 'Paragraph',
-          },
-        ],
-      },
-    ]
-    const config = {
-      ...slateToDomConfig,
-      elementTransforms: {
-        ...slateToDomConfig.elementTransforms,
-        p: ({ node, children }: { node?: any; children?: ChildNode[] }) => {
-          return new Element(
-            'p',
-            {
-              style: transformStyleObjectToString(node.style),
-            },
-            children,
-          )
-        },
-      },
-    }
-    expect(slateToHtml(slate, config)).toEqual(html)
-  })
-
-  it('mark transforms', () => {
-    const html =
-      '<p><span style="font-size: 96px; --text-color: #DD3A0A; @media screen { z-index: 1; color: var(--text-color) }"><strong>Paragraph</strong></span></p>'
-    const slate = [
-      {
-        type: 'p',
-        children: [
-          {
-            bold: true,
-            style: {
-              fontSize: '96px',
-              '--text-color': '#DD3A0A',
-              '@media screen': {
-                zIndex: '1',
-                color: 'var(--text-color)',
-              },
-            },
-            text: 'Paragraph',
-          },
-        ],
-      },
-    ]
-    const config = {
-      ...slateToDomConfig,
-      markTransforms: {
-        ...slateToDomConfig.markTransforms,
-        style: ({ node }: { node?: any }) => {
-          return new Element('span', {
-            style: transformStyleObjectToString(node.style),
-          })
-        },
-      },
-    }
-    expect(slateToHtml(slate, config)).toEqual(html)
-  })
-
-  it('mark transforms on multiple marks', () => {
-    const html =
-      '<p>This is editable <span style="font-size: 20px; font-weight: 600; text-decoration: underline dotted"><strong>rich</strong></span> text, <span style="text-decoration: underline"><i>much</i></span> better than a <span style="color: red"><pre><code>&lt;textarea&gt;</code></pre></span>!</p>'
-    const slate = [
-      {
-        type: 'p',
-        children: [
-          {
-            text: 'This is editable ',
-          },
-          {
-            text: 'rich',
-            bold: true,
-            style: {
-              fontSize: '20px',
-              fontWeight: 600,
-              textDecoration: 'underline dotted',
-            },
-          },
-          {
-            text: ' text, ',
-          },
-          {
-            text: 'much',
-            italic: true,
-            style: {
-              textDecoration: 'underline',
-            },
-          },
-          {
-            text: ' better than a ',
-          },
-          {
-            text: '<textarea>',
-            code: true,
-            style: {
-              color: 'red',
-            },
-          },
-          {
-            text: '!',
-          },
-        ],
-      },
-    ]
-    const config = {
-      ...slateToDomConfig,
-      markTransforms: {
-        ...slateToDomConfig.markTransforms,
-        style: ({ node }: { node?: any }) => {
-          return new Element('span', {
-            style: transformStyleObjectToString(node.style),
           })
         },
       },
